requirements.txt
python-dotenv
requests
requests-toolbelt
regex
